Revert "compile"
This reverts commit 560e78b0e66d48346057780caa97e1a698d68dfb.

+++ bchain/baseparser.go
@@ -469,7 +469,7 @@ func (p *BaseParser) UnpackVarBytes(buf []byte) ([]byte, int) {
 }
 
 func (p *BaseParser) PackVarBytes(bufValue []byte, buf []byte, varBuf []byte) []byte {
-	l := p.PackVarint(len(bufValue), varBuf)
+	l := p.PackVarint(uint(len(bufValue)), varBuf)
 	buf = append(buf, varBuf[:l]...)
 	buf = append(buf, bufValue...)
 	return buf
